finish implementing trending playlists carousel using lineups
