+++ packages/mobile/src/screens/explore-screen/components/TrendingPlaylists.tsx
@@ -1,21 +1,22 @@
-import React, { useMemo } from 'react'
+import React from 'react'
 
 import { useTrendingPlaylists } from '@audius/common/api'
 import { exploreMessages as messages } from '@audius/common/messages'
+import { trendingPlaylistsPageLineupActions } from '@audius/common/store'
 
-import { CollectionTileCarousel } from './CollectionTileCarousel'
+import { CollectionLineupCarousel } from './CollectionLineupCarousel'
 import { ExploreSection } from './ExploreSection'
 
 export const TrendingPlaylists = () => {
-  const { data: trendingPlaylists, isLoading } = useTrendingPlaylists()
-
-  const playlistIds = useMemo(() => {
-    return trendingPlaylists.map(({ id }) => id)
-  }, [trendingPlaylists])
+  const { lineup, isLoading } = useTrendingPlaylists()
 
   return (
     <ExploreSection title={messages.trendingPlaylists} isLoading={isLoading}>
-      <CollectionTileCarousel collectionIds={playlistIds} isTrending />
+      <CollectionLineupCarousel
+        lineup={lineup}
+        isTrending
+        actions={trendingPlaylistsPageLineupActions}
+      />
     </ExploreSection>
   )
 }
